Enhance CourseTemplate management by implementing deletion restrictions and improving session handling.
- Added a check to prevent deletion of CourseTemplates that are associated with existing courses.
- Updated StoreCourseTemplateAction to set 'is_self_paced' based on the selected course type.
- Refactored media handling in StoreOpinionAction and UpdateOpinionAction to specify media types.
- Improved CourseRuner component to dynamically calculate end time based on session duration.
- Enhanced CourseUpdateOrCreate component to utilize CourseTemplate data for course creation and updates.
- Updated CourseTemplateUpdateOrCreate component to manage session types and levels more effectively.
- Refined UI components for better user experience in course and opinion management.

+++ app/Actions/CourseTemplate/DeleteCourseTemplateAction.php
@@ -17,6 +17,10 @@ class DeleteCourseTemplateAction
     public function handle(CourseTemplate $courseTemplate): bool
     {
         return DB::transaction(function () use ($courseTemplate) {
+            abort_if($courseTemplate->courses()->exists(), 403, trans('courseTemplate.exceptions.not_allowed_to_delete_course_template_due_to_courses'));
+
+            $courseTemplate->sessionTemplates()->delete();
+
             return $courseTemplate->delete();
         });
     }
